+++ shuttle/templates/driver_check_in/driver_check_in.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_bases/shuttle_base.html' %}
+{% extends 'shuttle_base.html' %}
 <!-- TODO: select next location after submitting so it is easier for drivers -->
 {% block body_content %}
 <title>Shuttle | {% block page_title %}Driver Check In{% endblock %}</title>
