+++ shuttle/templates/driver_logs/driver_logs.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_bases/shuttle_base.html' %}
+{% extends 'shuttle_base.html' %}
 <title>Shuttle | {% block page_title %}Driver Logs{% endblock %}</title>
 {% block body_content %}
     <h4>Please select a date below to view its corresponding logs</h4>
