+++ shuttle/templates/homepage/index.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_bases/shuttle_base.html' %}
+{% extends 'shuttle_base.html' %}
 <!-- TODO Make work for multiple drivers -->
 {% block body_content %}
