+++ shuttle/templates/request_shuttle/request_shuttle.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_bases/shuttle_base.html' %}
+{% extends 'shuttle_base.html' %}
 <!-- TODO: Maybe reset waitlist every night. Maybe make submenu to shorten dropdown -->
 <title>Shuttle | {% block page_title %}Request{% endblock %}</title>
 {% block body_content %}
