+++ shuttle/templates/schedules/edit_shuttle_schedule.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_bases/shuttle_base.html' %}
+{% extends 'shuttle_base.html' %}
     <title>Shuttle | {% block page_title %}Edit Schedule{% endblock %}</title>
 {% block body_content %}
