+++ shuttle/templates/shuttle_base.html
@@ -75,7 +75,7 @@
 
         {% endblock %}
 
-        {% block nav %}{% include 'shuttle_bases/nav.html' %}{% endblock %}
+        {% block nav %}{% include 'nav.html' %}{% endblock %}
 
          <div class="nav margins">{% block subnav %}{% endblock %}</div>
