+++ shuttle/templates/users/users.html
@@ -1,4 +1,4 @@
-{% extends 'shuttle_bases/shuttle_base.html' %}
+{% extends 'shuttle_base.html' %}
     <title>Shuttle | {% block page_title %}Users{% endblock %}</title>
 
 {% block body_content %}
