+++ shuttle/views.py
@@ -2,18 +2,18 @@ from flask import render_template, make_response, redirect
 from flask import session as flask_session
 from flask_classy import FlaskView, route
 from shuttle import app
-from shuttle.homepage.homepage_controller import HomePageController
+from shuttle.shuttle_controller import ShuttleController
 
 
 class View(FlaskView):
     def __init__(self):
-        self.hc = HomePageController()
+        self.sc = ShuttleController()
 
     @route('/')
     def index(self):
-        check_in_data = self.hc.grab_check_in_driver_data()
-        route_data = self.hc.grab_current_route()
-        return render_template('shuttle_bases/index.html', **locals())
+        check_in_data = self.sc.grab_check_in_driver_data()
+        route_data = self.sc.grab_current_route()
+        return render_template('homepage/index.html', **locals())
 
     @route('/clear')
     def clear_session(self):
